+++ inst.ml
@@ -9,4 +9,4 @@ let () =
   Printf.printf "  let data_path = \"%s/imaplet\"\n%!" data;
   Printf.printf "  let users_path = \"%s/imaplet/users\"\n%!" data;
   Printf.printf "  let config_path = \"%s/imaplet/imaplet.cf\"\n%!" data;
-  Printf.printf "  let lmtp_srv_exec = \"%s/imaplet_lmtp\"\n%!" bin;
+  Printf.printf "  let smtp_srv_exec = \"%s/imaplet_smtp\"\n%!" bin;