+++ lib/commands/account.ml
@@ -53,10 +53,13 @@ let parse_users buff user password =
   with _ ->
     false
 
+let b64decode b64 =
+   (*let buff = Str.global_replace (Str.regexp "=$") "" b64 in*)
+   Cstruct.to_string (Nocrypto.Base64.decode (Cstruct.of_string b64))
+
 let parse_user_b64 b64 =
   try 
-   let buff = Str.global_replace (Str.regexp "=$") "" b64 in
-   let buff = Cstruct.to_string (Nocrypto.Base64.decode (Cstruct.of_string buff)) in (** need to log this if it fails **)
+   let buff = b64decode b64 in (** need to log this if it fails **)
    let _ = Str.search_forward (Str.regexp
    "^\\([^\\]+\\)\000\\([^\\]+\\)\000\\([^\\]+\\)$") buff 0 in
    let u1 = Str.matched_group 1 buff in
@@ -79,7 +82,9 @@ let rec read_users r user password =
     function 
       | Some res -> 
         if match_user res user then (
-          if parse_users res user password then
+          if password = None then
+            return (true)
+          else if parse_users res user (option_value_exn password) then
             return (true)
           else
             return (false)
@@ -88,12 +93,20 @@ let rec read_users r user password =
       | None -> return (false)
 
 (** have to make users configurable **)
-let authenticate_user ?(users=Install.users_path) user password =
+let authenticate_user ?(b64=false) ?(users=Install.users_path) user ?password () =
+  let (user,password) =
+    if b64 = false then
+      (user,password)
+    else (
+      match password with 
+      | None -> (b64decode user, None)
+      | Some p -> (b64decode user, Some (b64decode p))
+    )
+  in
   Lwt_io.with_file ~mode:Lwt_io.Input users (fun r -> read_users r user password)
 
-
 let auth_user user password resp_ok resp_no =
-  authenticate_user user password >>= fun res ->
+  authenticate_user user ~password () >>= fun res ->
   if res then
     return (`Ok (Resp_Ok
     (None,Utils.formated_capability(Configuration.auth_capability)), user, password))
@@ -102,13 +115,18 @@ let auth_user user password resp_ok resp_no =
 
 let plain_auth text =
   match (parse_user_b64 text) with
+  | Some (u,p) -> authenticate_user u ?password:(Some p) ()
+  | None -> return false
+
+let _plain_auth text =
+  match (parse_user_b64 text) with
   | Some (u,p) -> auth_user u p "AUTHENTICATE" "PASSWORD"
   | None -> return (`Error (Resp_No (None,"PASSWORD")))
 
 (** TBD authenticate plain against users file **)
 let authenticate auth_type text =
   match auth_type with 
-    | Auth_Plain -> plain_auth text
+    | Auth_Plain -> _plain_auth text
     | _ -> return (`Error (Resp_No (None,"Authentication Type")))
 
 (** TBD **)