+++ lib/commands/install.ml
@@ -1,4 +1,4 @@
   let data_path = "/usr/local/share/imaplet"
   let users_path = "/usr/local/share/imaplet/users"
   let config_path = "/usr/local/share/imaplet/imaplet.cf"
-  let lmtp_srv_exec = "/usr/local/bin/imaplet_lmtp"
+  let smtp_srv_exec = "/usr/local/bin/imaplet_smtp"